# docs/swagger.yaml
basePath: /
definitions:
...
  episodes.CreateClipRequest:
    properties:
      end_time:
        example: 45
        type: number
      label:
        example: advertisement
        minLength: 1
        type: string
      start_time:
        example: 30
        minimum: 0
        type: number
    required:
    - end_time
    - label
    type: object
  episodes.EpisodeClipResponse:
    properties:
      approved:
        example: true
        type: boolean
      auto_labeled:
        example: false
        type: boolean
      created_at:
        example: "2025-10-02T13:00:00Z"
        type: string
      duration:
        example: 15
        type: number
      error_message:
        example: ""
        type: string
      extracted:
        example: false
        type: boolean
      filename:
        example: clip_a1b2c3d4.wav
        type: string
      label:
        example: advertisement
        type: string
      label_confidence:
        example: 0.95
        type: number
      label_method:
        enum:
        - manual
        - peak_detection
        example: manual
        type: string
      original_end_time:
        example: 45
        type: number
      original_start_time:
        example: 30
        type: number
      size_bytes:
        example: 480332
        type: integer
      status:
        enum:
        - detected
        - queued
        - processing
        - ready
        - failed
        example: queued
        type: string
      updated_at:
        example: "2025-10-02T13:00:00Z"
        type: string
      uuid:
        example: a1b2c3d4-e5f6-4a7b-8c9d-0e1f2a3b4c5d
        type: string
    type: object
...
  episodes.UpdateLabelRequest:
    properties:
      label:
        example: music
        minLength: 1
        type: string
    required:
    - label
    type: object
...
  types.ErrorResponse:
    properties:
      details:
        description: Additional error details
      error:
        description: Error code/type
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@example.com
    name: Podcast Player API Support
    url: https://github.com/killallgit/killallplayer-api
  description: |-
    A REST API for podcast discovery and episode management.
    This API provides endpoints for searching podcasts, managing episodes,
    and generating waveforms for audio visualization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Podcast Player API
  version: "1.0"
paths:
...
  /api/v1/episodes/{id}/clips:
    get:
      description: Get all clips created for this episode with optional status and
        approval filters
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by status
        enum:
        - queued
        - processing
        - ready
        - failed
        - detected
        in: query
        name: status
        type: string
      - description: Filter by approval status (true/false)
        in: query
        name: approved
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/episodes.EpisodeClipResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List clips for episode
      tags:
      - episodes
    post:
      consumes:
      - application/json
      description: Create a new audio clip from this episode at the specified time
        range. Manual clips are automatically approved and will be extracted during
        dataset export.
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip creation parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/episodes.CreateClipRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Clip created successfully (approved=true, status=pending)
          schema:
            $ref: '#/definitions/episodes.EpisodeClipResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create clip for episode
      tags:
      - episodes
  /api/v1/episodes/{id}/clips/{uuid}:
    delete:
      description: Delete a clip and its audio file from this episode
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip UUID
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete clip
      tags:
      - episodes
    get:
      description: Get details of a specific clip for this episode
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/episodes.EpisodeClipResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get clip details
      tags:
      - episodes
  /api/v1/episodes/{id}/clips/{uuid}/approve:
    put:
      description: Mark a clip as approved for extraction. This is used for clips
        created by analysis (status=detected, approved=false) to trigger audio extraction.
        Sets approved=true and queues the clip for processing.
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Clip approved and queued for extraction
          schema:
            $ref: '#/definitions/episodes.EpisodeClipResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Approve clip for extraction
      tags:
      - episodes
  /api/v1/episodes/{id}/clips/{uuid}/label:
    put:
      consumes:
      - application/json
      description: Update the classification label for a clip
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: New label
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/episodes.UpdateLabelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/episodes.EpisodeClipResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update clip label
      tags:
      - episodes
...
schemes:
- http
- https
swagger: "2.0"
...